Refactor webapi routing structure
